better visualization

+++ include/MapPoint.cc
@@ -25,23 +25,23 @@ cv::Mat MapPoint::getNormal() {
     return mNormalVector.clone();
 }
 
-void MapPoint::setNormalDirection(double phi, double psi) {
+void MapPoint::setNormalDirection(float phi, float psi) {
     lock_guard<mutex> lock(mMutexPos);
     mNormalDirection[0] = phi;
     mNormalDirection[1] = psi;
-    double data[] = {cos(phi) * sin(psi), sin(phi) * sin(psi), cos(psi)};
-    cv::Mat(3, 1, CV_64F, data).copyTo(mNormalVector);
+    float data[] = {cos(phi) * sin(psi), sin(phi) * sin(psi), cos(psi)};
+    cv::Mat(3, 1, CV_32F, data).copyTo(mNormalVector);
     Eigen::Vector3d z;
     z << 0, 0, 1;
     Eigen::Vector3d nw;
     nw << cos(phi) * sin(psi), sin(phi) * sin(psi), cos(psi);
     Eigen::Vector3d v = (-nw).cross(z);
-    double c = -z.dot(nw);
+    float c = -z.dot(nw);
     Eigen::Matrix3d Kn = ev::skew(v);
     Rn = Eigen::Matrix3d::Identity() + Kn + Kn * Kn / (1 + c);
 }
 
-std::array<double, 2>& MapPoint::getNormalDirection() {
+std::array<float, 2>& MapPoint::getNormalDirection() {
     lock_guard<mutex> lock(mMutexPos);
     return mNormalDirection;
 }
