+++ include/MapPoint.h
@@ -27,8 +27,8 @@ public:
     cv::Mat getNormal();
     // not implemented
     void setNormal();
-    std::array<double, 2>& getNormalDirection();
-    void setNormalDirection(double phi, double psi);
+    std::array<float, 2>& getNormalDirection();
+    void setNormalDirection(float phi, float psi);
 
     void addObservation(shared_ptr<KeyFrame>& pKF);
 
@@ -38,7 +38,7 @@ public:
 public:
     unsigned int mnId;
     // Normal vector of the plane
-    std::array<double, 2> mNormalDirection;
+    std::array<float, 2> mNormalDirection;
     cv::Mat mNormalVector;
 
     // Synthetic image of the plane
@@ -61,7 +61,7 @@ public:
     cv::Mat mWorldPos;
 
     // inverse depth to the origin
-    double d;
+    float d;
 
     mutex mMutexPos;
